+++ src/main/java/phoupraw/mcmod/trifleclient/mixin/minecraft/MAbstractClientPlayerEntity.java
@@ -1,0 +1,15 @@
+package phoupraw.mcmod.trifleclient.mixin.minecraft;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import net.minecraft.client.network.AbstractClientPlayerEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import phoupraw.mcmod.trifleclient.config.TCConfigs;
+
+@Mixin(AbstractClientPlayerEntity.class)
+abstract class MAbstractClientPlayerEntity {
+    @ModifyExpressionValue(method = "getFovMultiplier", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/network/AbstractClientPlayerEntity;getAttributeValue(Lnet/minecraft/registry/entry/RegistryEntry;)D"))
+    private double limitSpeed(double original) {
+        return Math.min(original, TCConfigs.A().getLimitSpeed());
+    }
+}