+++ CHANGELOG.md
@@ -299,5 +299,8 @@
 ## 修复
 - 不自动收割稻米
 #
+## 新增
+- 限制速度也会限制FOV
+- `/trifleclient find blocks <condition>`
 ## 修复
 - 自动收割补种有时消耗两个种子（[#12](https://github.com/Phoupraw/TrifleClient/issues/12)）